+++ lib/decorators/array.js
