+++ lib/decorators/flow.js
@@ -5,7 +5,7 @@
 (function(define) { 'use strict';
 define(function(require) {
 
-	var setTimer = require('./timer').set;
+	var setTimer = require('../timer').set;
 
 	return function flow(Promise) {
