+++ lib/decorators/inspect.js
