+++ lib/decorators/iterate.js
