+++ lib/decorators/timed.js
@@ -5,7 +5,7 @@
 (function(define) { 'use strict';
 define(function(require) {
 
-	var timer = require('./timer');
+	var timer = require('../timer');
 
 	return function timed(Promise) {
 		/**
