+++ lib/decorators/with.js
